+++ src/Command/Manager.php
@@ -79,7 +79,7 @@ class Manager
         $this->registerCommand('install',       'OmekaCli\Command\InstallCommand');
         $this->registerCommand('list',          'OmekaCli\Command\ListCommand');
         $this->registerCommand('options',       'OmekaCli\Command\OptionsCommand');
-        $this->registerCommand('plugin',        'OmekaCli\Command\PluginCommand');
+        $this->registerCommand('plugin',        'OmekaCli\Command\PluginCommands\PluginCommand');
         $this->registerCommand('snapshot',      'OmekaCli\Command\SnapshotCommand', array('snap'));
         $this->registerCommand('upgrade',       'OmekaCli\Command\UpgradeCommand');
         $this->registerCommand('version',       'OmekaCli\Command\VersionCommand');