+++ CHANGELOG.md
@@ -5,11 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.17.0] - 2017-08-08
 ### Changed
 - `install` command installs last tagged version by default.
 
 ### Added
 - `upgrade` command.
+- `snapshot` command.
 
 ### Fixed
 - Minor fixes.